refactor: remvoved prelude and filtered entries. Disabled fast-near for now
