+++ examples/account_key_pooling.rs
@@ -3,8 +3,9 @@
 ///
 /// This is an example of how to use account key pooling to send multiple transactions
 /// using different keys.
-use near_api::prelude::*;
+use near_api::*;
 use near_token::NearToken;
+use signer::generate_secret_key;
 
 use std::sync::Arc;
