+++ examples/deploy_and_call_method.rs
@@ -1,4 +1,4 @@
-use near_api::{prelude::*, types::Data};
+use near_api::*;
 
 #[tokio::main]
 async fn main() {
