+++ src/tokens.rs
@@ -23,13 +23,13 @@ use crate::{
     },
     contract::Contract,
     errors::{BuilderError, FTValidatorError, ValidationError},
-    prelude::*,
     transactions::{ConstructTransaction, TransactionWithSign},
     types::{
         tokens::{FTBalance, UserBalance},
         transactions::PrepopulateTransaction,
         Data,
     },
+    NetworkConfig, StorageDeposit,
 };
 
 type Result<T> = core::result::Result<T, BuilderError>;
